requirements.txt
pyyaml
flake8
mypy
markdown-it-py
